backend/requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0
psycopg2-binary
alembic
pydantic
python-dotenv
